+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - You can now properly create a PC from the "New Tab" page
 - Formatting (new lines in particular) is now preserved in descriptions when creating entries with user-entered text.
 - Editing a description no longer suggests creating references to entries/documents that can't be related to that entry 
+- Dragging an entry, session, campaign, or setting from the tree properly inserts a link to it in the editor again
 
 ## v1.4.0 - Now in Russian
 ![](https://img.shields.io/badge/release%20date-September%2022%2C%202025-blue)